Handle data serialized on Windows
Before this commit, serialization used the array module, but the size of
units in that varies depending on the host OS. As a result data
serialized on Windows would use 32-bit units instead of 64-bit ones.

This commit uses struct for serialization instead of array, eliminating
the ambiguity issue. Additionally, it detects if the serialized data
used the Windows size, and loads it in that size if necessary, so old
data will still work. (Even if the data was serialized into smaller
containers on Windows no data was lost due to a separate issue, which
will be fixed in a further commit.)

Note that this issue only affected the serialization code, and in memory
containers were consistent on platforms already.