Fix math issues
This fixes math issues with bloom filters so that they take up the
proper amount of memory instead of 8x too much. This issue stemmed from
confusion around which values were in bits and which in bytes.

It also adds a header and version tag to serialized data, to
differentiate data created with new math from data created with old
math. Deserialization should load old data but the tests for this aren't
written yet.