Disallow invalid args for bloom filter
Includes a test to check this works.

While working on the backwards compatability, it came up that in current
releases of preshed it's possible to specify 0 for the size or hash
function count. If the hcount is 0 the add/contains functions do
nothing, but do not fail. If size is 0 and hcount is not you'll get a
core dump.