Fix bloom size issues with Windows serialization and in general (#38)
* Handle data serialized on Windows

Before this commit, serialization used the array module, but the size of
units in that varies depending on the host OS. As a result data
serialized on Windows would use 32-bit units instead of 64-bit ones.

This commit uses struct for serialization instead of array, eliminating
the ambiguity issue. Additionally, it detects if the serialized data
used the Windows size, and loads it in that size if necessary, so old
data will still work. (Even if the data was serialized into smaller
containers on Windows no data was lost due to a separate issue, which
will be fixed in a further commit.)

Note that this issue only affected the serialization code, and in memory
containers were consistent on platforms already.

* Fix math issues

This fixes math issues with bloom filters so that they take up the
proper amount of memory instead of 8x too much. This issue stemmed from
confusion around which values were in bits and which in bytes.

It also adds a header and version tag to serialized data, to
differentiate data created with new math from data created with old
math. Deserialization should load old data but the tests for this aren't
written yet.

* Add tests for legacy, Linux one works

Windows one still needs a little work

* Formatting of tests

* Fix Windows

This is pretty complicated! The tests seem OK though.

The sample Windows bytes were generated on Linux, I need to check that
they are actually like that on Windows.

* Disallow invalid args for bloom filter

Includes a test to check this works.

While working on the backwards compatability, it came up that in current
releases of preshed it's possible to specify 0 for the size or hash
function count. If the hcount is 0 the add/contains functions do
nothing, but do not fail. If size is 0 and hcount is not you'll get a
core dump.

* Fix the 8.

The 8 here is "the size of a byte in bits", not the size of the key
(which happens to be the same value). Math...

* Remove comment

* Add checks to avoid alloc'ing with len 0

* Temporarily skip windows test to check for segfault cause

* Skip other legacy test

* Turn tests back on, but make legacy decoding a no-op

Expect the tests to fail but without a segfault.

* Turn on legacy decoding, but disable all Windows code

Expect the tests to pass but segfault, given the non-windows legacy test
segfaulted before.

* Fix segfault

The issue was that there was no check that length was a multiple of
KEY_BITS. Based on the init code, that should always be the case, but
that shouldn't be relied on.

* Put Windows code back, avoid overflow with seed value

* Fix buflen calculation

Objects made with the current init code will always have a bitfield size
that's a multiple of container size. But if data is deserialized from
legacy data the old size may not be a multiple, so that should be
handled too.

* Test re-serializing legacy data

* Clean up ceiling division

math.ceil behaves a little unpredictably because sometimes normal
division can give a float and sometimes it doesn't with Cython types.
I'm not sure this is the best way, but relying strictly on integer
division avoids any magic with floating point numbers and lets the math
be consistent.

* Use anonymous enum for constants, change version to constant

* Feedback from code review

* Run tests in dev mode

* Suggestions from code review

- use constants where possible
- make decode_len calculation uniform

* Remove overly-complicated int truncation

* Remove python 3.6 tests

* Store seed as L, not Q

* More efficient buffer construction

* Use memcpy instead of a for loop

* Update preshed/bloom.pyx

Co-authored-by: Daniël de Kok <me@github.danieldk.eu>

* Change version mismatch to raise exception

* significant -> "actually used"

---------

Co-authored-by: Daniël de Kok <me@github.danieldk.eu>